fix: apply review feedback for compact-cli-dev:core skill
- Add WALLET_NOT_FOUND, STALE_UTXO, INVALID_SEED classification to
  classifyError() (was enum-only, fell through to UNKNOWN)
- Extract findComposeFile() to shared lib/devnet.ts, import from both
  devnet/start.ts and devnet/stop.ts
- Replace hard-coded throttleTime(5_000) with DUST_POLL_INTERVAL constant
  in balance.ts and dust/status.ts
- Improve skill description with purpose clause and specific trigger phrases
- Document template placeholders and add templates/ pointer to SKILL.md
