+++ plugins/compact-cli-dev/skills/core/templates/cli/src/commands/dust/status.ts
@@ -1,6 +1,7 @@
 import { Args } from "@oclif/core";
 import * as Rx from "rxjs";
 import { BaseCommand } from "../../base-command.js";
+import { DUST_POLL_INTERVAL } from "../../lib/constants.js";
 import { buildFacade, getWallet } from "../../lib/wallet.js";
 import { withSpinner } from "../../lib/progress.js";
 
@@ -19,7 +20,7 @@ export default class DustStatus extends BaseCommand {
 		try {
 			const state = await Rx.firstValueFrom(
 				ctx.facade.state().pipe(
-					Rx.throttleTime(5_000),
+					Rx.throttleTime(DUST_POLL_INTERVAL),
 					Rx.filter((s) => s.isSynced),
 				),
 			);
